+++ lib/pages/demo02.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class DemoApp02 extends StatelessWidget {
+  const DemoApp02({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Second Route'),
+      ),
+      body: Center(
+        child: ElevatedButton(
+          onPressed: () {
+            Navigator.pop(context);
+          },
+          child: const Text('Go back!'),
+        ),
+      ),
+    );
+  }
+}
